feat: migrating local database completely by SQLDelight

+++ feature/gratefulness/impl/src/commonMain/kotlin/com/joohnq/gratefulness/impl/ui/presentation/component/GratefulnessHistoryCard.kt
@@ -25,7 +25,7 @@ import org.jetbrains.compose.resources.stringResource
 fun GratefulnessHistoryCard(
     modifier: Modifier = Modifier,
     gratefulness: Gratefulness,
-    onDelete: (Int) -> Unit,
+    onDelete: (Long) -> Unit,
 ) {
     CardWithMoreMenuLayout(
         modifier = modifier,
