+++ feature/self-journal/api/src/commonMain/kotlin/com/joohnq/self_journal/api/entity/SelfJournalRecord.kt
@@ -5,7 +5,7 @@ import com.joohnq.mood.api.entity.Mood
 import kotlinx.datetime.LocalDateTime
 
 data class SelfJournalRecord(
-    val id: Int = -1,
+    val id: Long = -1,
     val mood: Mood = Mood.Neutral,
     val title: String = "",
     val description: String = "",
