+++ feature/self-journal/api/src/commonMain/kotlin/com/joohnq/self_journal/api/use_case/GetSelfJournalByIdUseCase.kt
@@ -6,7 +6,7 @@ import com.joohnq.self_journal.api.repository.SelfJournalRepository
 class GetSelfJournalByIdUseCase(
     private val repository: SelfJournalRepository,
 ) {
-    suspend operator fun invoke(id: Int): Result<SelfJournalRecord> =
+    suspend operator fun invoke(id: Long): Result<SelfJournalRecord> =
         try {
             val selfJournal = repository.getById(id)
