+++ feature/self-journal/history/src/commonMain/kotlin/com/joohnq/self_journal/presentation/SelfJournalHistoryViewModel.kt
@@ -49,7 +49,7 @@ class SelfJournalHistoryViewModel(
             }.launchIn(viewModelScope)
     }
 
-    private fun delete(id: Int) {
+    private fun delete(id: Long) {
         viewModelScope.launch {
             try {
                 deleteSelfJournalsUseCase(id).getOrThrow()
