+++ feature/self-journal/overview/src/commonMain/kotlin/com/joohnq/self_journal/presentation/SelfJournalOverviewScreen.kt
@@ -10,7 +10,7 @@ fun SelfJournalOverviewScreen(
     onNavigateAddSelfJournal: () -> Unit,
     onNavigateToSelfJournalHistory: () -> Unit,
     onGoBack: () -> Unit,
-    onEditSelfJournal: (Int) -> Unit,
+    onEditSelfJournal: (Long) -> Unit,
     viewModel: SelfJournalOverviewViewModel = sharedViewModel(),
 ) {
     val state by viewModel.state.collectAsState()
