+++ feature/sleep-quality/history/src/commonMain/kotlin/com/joohnq/sleep_quality/history/presentation/SleepQualityHistoryViewModel.kt
@@ -47,7 +47,7 @@ class SleepQualityHistoryViewModel(
             }.launchIn(viewModelScope)
     }
 
-    private fun delete(id: Int) {
+    private fun delete(id: Long) {
         viewModelScope.launch {
             try {
                 deleteSleepQualityUseCase(id).getOrThrow()
