+++ feature/sleep-quality/overview/src/commonMain/kotlin/com/joohnq/sleep_quality/overview/presentation/SleepQualityOverviewViewModel.kt
@@ -47,7 +47,7 @@ class SleepQualityOverviewViewModel(
             }.launchIn(viewModelScope)
     }
 
-    private fun delete(id: Int) {
+    private fun delete(id: Long) {
         viewModelScope.launch {
             try {
                 deleteSleepQualityUseCase(id).getOrThrow()
